+++ cinetica/relativo/__init__.py
@@ -1,0 +1,1 @@
+from .velocidad_relativa import *